conf: Extract BaseConfig interface
Start to break apart abstraction and interface as it locks me down too
much. This change will allow creation of a much more flexible config
implementation internally while not exposing too much of that to the
outside. Will be apparant in following commits.

Relates to #1138
