+++ pyatv/conf.py
@@ -7,13 +7,12 @@ provide configurations for you.
 from ipaddress import IPv4Address
 from typing import Dict, List, Mapping, Optional
 
-from pyatv import exceptions
 from pyatv.const import Protocol
-from pyatv.interface import BaseService, DeviceInfo
+from pyatv.interface import BaseConfig, BaseService, DeviceInfo
 
 
-class AppleTV:
-    """Representation of an Apple TV configuration.
+class AppleTV(BaseConfig):
+    """Representation of a device configuration.
 
     An instance of this class represents a single device. A device can have
     several services depending on the protocols it supports, e.g. DMAP or
@@ -29,12 +28,12 @@ class AppleTV:
         device_info: Optional[DeviceInfo] = None,
     ) -> None:
         """Initialize a new AppleTV."""
+        super().__init__(properties or {})
         self._address = address
         self._name = name
         self._deep_sleep = deep_sleep
         self._services: Dict[Protocol, BaseService] = {}
         self._device_info = device_info or DeviceInfo({})
-        self.properties: Mapping[str, Mapping[str, str]] = properties or {}
 
     @property
     def address(self) -> IPv4Address:
@@ -50,28 +49,6 @@ class AppleTV:
     def deep_sleep(self) -> bool:
         """If device is in deep sleep."""
         return self._deep_sleep
-
-    @property
-    def ready(self) -> bool:
-        """Return if configuration is ready, (at least one service with identifier)."""
-        for service in self.services:
-            if service.identifier:
-                return True
-        return False
-
-    @property
-    def identifier(self) -> Optional[str]:
-        """Return the main identifier associated with this device."""
-        for prot in [Protocol.MRP, Protocol.DMAP, Protocol.AirPlay, Protocol.RAOP]:
-            service = self._services.get(prot)
-            if service:
-                return service.identifier
-        return None
-
-    @property
-    def all_identifiers(self) -> List[str]:
-        """Return all unique identifiers for this device."""
-        return [x.identifier for x in self.services if x.identifier is not None]
 
     def add_service(self, service: BaseService) -> None:
         """Add a new service.
@@ -97,56 +74,10 @@ class AppleTV:
         """Return all supported services."""
         return list(self._services.values())
 
-    def main_service(self, protocol: Optional[Protocol] = None) -> BaseService:
-        """Return suggested service used to establish connection."""
-        protocols = (
-            [protocol]
-            if protocol is not None
-            else [Protocol.MRP, Protocol.DMAP, Protocol.AirPlay, Protocol.RAOP]
-        )
-
-        for prot in protocols:
-            service = self._services.get(prot)
-            if service is not None:
-                return service
-
-        raise exceptions.NoServiceError("no service to connect to")
-
-    def set_credentials(self, protocol: Protocol, credentials: str) -> bool:
-        """Set credentials for a protocol if it exists."""
-        service = self.get_service(protocol)
-        if service:
-            service.credentials = credentials
-            return True
-        return False
-
     @property
     def device_info(self) -> DeviceInfo:
         """Return general device information."""
         return self._device_info
-
-    def __eq__(self, other) -> bool:
-        """Compare instance with another instance."""
-        if isinstance(other, self.__class__):
-            return self.identifier == other.identifier
-        return False
-
-    def __str__(self) -> str:
-        """Return a string representation of this object."""
-        device_info = self.device_info
-        services = "\n".join([f" - {s}" for s in self._services.values()])
-        identifiers = "\n".join([f" - {x}" for x in self.all_identifiers])
-        return (
-            f"       Name: {self.name}\n"
-            f"   Model/SW: {device_info}\n"
-            f"    Address: {self.address}\n"
-            f"        MAC: {self.device_info.mac}\n"
-            f" Deep Sleep: {self.deep_sleep}\n"
-            f"Identifiers:\n"
-            f"{identifiers}\n"
-            f"Services:\n"
-            f"{services}"
-        )
 
 
 # pylint: disable=too-few-public-methods
