+++ pyatv/protocols/airplay/pairing.py
@@ -2,9 +2,9 @@
 import logging
 from typing import Optional
 
-from pyatv import conf, exceptions
+from pyatv import exceptions
 from pyatv.auth.hap_pairing import PairSetupProcedure
-from pyatv.interface import BaseService, PairingHandler
+from pyatv.interface import BaseConfig, BaseService, PairingHandler
 from pyatv.protocols.airplay.auth import AuthenticationType, pair_setup
 from pyatv.protocols.airplay.features import AirPlayFlags, parse
 from pyatv.support import error_handler
@@ -28,7 +28,7 @@ class AirPlayPairingHandler(PairingHandler):
 
     def __init__(
         self,
-        config: conf.AppleTV,
+        config: BaseConfig,
         service: BaseService,
         session_manager: ClientSessionManager,
         auth_type: AuthenticationType,
