+++ pyatv/protocols/companion/__init__.py
@@ -8,7 +8,6 @@ from typing import Any, Dict, Generator, List, Mapping, Optional, Set, cast
 
 from pyatv import conf, exceptions
 from pyatv.auth.hap_srp import SRPAuthHandler
-from pyatv.conf import AppleTV
 from pyatv.const import DeviceModel, FeatureName, FeatureState, InputAction, Protocol
 from pyatv.core import SetupData, mdns
 from pyatv.core.device_info import lookup_model
@@ -17,6 +16,7 @@ from pyatv.interface import (
     App,
     Apps,
     Audio,
+    BaseConfig,
     BaseService,
     DeviceInfo,
     FeatureInfo,
@@ -90,7 +90,7 @@ class HidCommand(Enum):
 class CompanionAPI(CompanionConnectionListener):
     """Implementation of Companion API."""
 
-    def __init__(self, config: AppleTV, loop: asyncio.AbstractEventLoop):
+    def __init__(self, config: BaseConfig, loop: asyncio.AbstractEventLoop):
         """Initialize a new CompanionAPI instance."""
         self.config = config
         self.loop = loop
@@ -346,7 +346,7 @@ def device_info(properties: Mapping[str, Any]) -> Dict[str, Any]:
 
 def setup(
     loop: asyncio.AbstractEventLoop,
-    config: conf.AppleTV,
+    config: BaseConfig,
     service: BaseService,
     device_listener: StateProducer,
     session_manager: ClientSessionManager,
@@ -386,7 +386,7 @@ def setup(
 
 
 def pair(
-    config: AppleTV,
+    config: BaseConfig,
     service: BaseService,
     session_manager: ClientSessionManager,
     loop: asyncio.AbstractEventLoop,
